+++ sqlservice/__init__.py
@@ -13,5 +13,4 @@ from .__pkg__ import (
 from .client import SQLClient
 from .core import destroy, make_identity, save, transaction
 from .model import ModelBase, declarative_base
-from .query import Query
-from .service import SQLService
+from .query import SQLQuery